#61 a more simple hierarchy
